music: split backend implementations into per-platform files
Put each MusicBackend implementation in its own translation unit
(musicbackenddesktop.cpp / musicbackendemscripten.cpp) instead of one
combined musicbackend.cpp. Each file carries a single top-level platform
guard and defines its own MusicBackend::create(); only one is active per
build, so no separate factory file or per-backend headers are needed.
